% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{exampleDataGraph}
\alias{exampleDataGraph}
\title{A simulated toy dataset that includes 3 igraph objects}
\format{A list of 3 igraph objects}
\usage{
data(exampleDataGraph)
}
\description{
(first one being the shared graph and second and third being task specific 1 and 2 graphs)
The graphs are generated from two related random graphs and the underlaying high dimensional gaussian distribution
generates the exampleData dataset. exampleDataGraph serves as a groundtruth to compare in demo(synthetic).
}
\keyword{datasets}
