% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{nip_37_data}
\alias{nip_37_data}
\title{NIPS word count dataset}
\format{a list of two nonnegative integer matrices (1:2900, 1:37) and (1:2911,1:37)
Columns are named with year_paperid and rows are names with word name}
\usage{
data(nip_37_data)
}
\description{
This NIPS Conference Papers 1987-2015 Data set is avaiable at UCI Machine Learning Repository.
The original dataset is in the form of a 11463 x 5812 matrix of word counts (11463 words and 5812 conference papers)
Due to the size of the original dataset, it is preprocessed and reduced to a list of two matrices (2900 x 37 and 2911 x 37)
The dataset consists of two tasks (early (up to 2006) and recent (after 2006) NIPS conference papers) with 37 words
}
\references{
'Poisson Random Fields for Dynamic Feature Models'. Perrone V., Jenkins P. A., Spano D., Teh Y. W. (2016)
}
\keyword{datasets}
